consolidate usps extensions shallow service integrations
